InfoFileRestored

+++ test/test_fileService.py
@@ -15,7 +15,8 @@ class PersonaService_FileManagementTests(unittest.TestCase):
     def eraseFiles(self):
         try:
             for fileName in os.listdir(self.folderPath):
-                os.remove(self.folderPath+self.pathFormat+fileName)
+                if '.txt' in fileName:
+                    os.remove(self.folderPath+self.pathFormat+fileName)
         except FileNotFoundError:
             exit
